% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotluck.options}
\alias{plotluck.options}
\title{Create options structure for \code{plotluck}}
\usage{
plotluck.options(...)
}
\arguments{
\item{...}{parameters to override default settings}
}
\value{
a named list of options, usable by function \code{plotluck}

\code{plotluck} accepts a list of options to modify its behavior. Calling
\code{plotluck.options} without arguments produces a list with the default
values. Specifying any number of attribute/value pairs overrides these
values, but leaves other options the same.

Available options and their meaning is described inline in the documentation
of \code{link{plotluck}}.
}
\description{
Create options structure for \code{plotluck}
}
\note{
\code{plotluck}'s aim is to provide a function that is usable
 "out-of-the-box", with no or very little manual tweaking. If you find
 yourself needing to change option values repeatedly or find the presets to
 be suboptimal, please contact the author.
}
\examples{
# list all default options
plotluck.options()

data(iris)
# default with violin plot
plotluck(iris, Species, Petal.Length)

# use box-and-whiskers plot instead
plotluck(iris, Species, Petal.Length, opts=plotluck.options(use.geom.violin=FALSE))
}
\seealso{
\code{\link{plotluck}}
}

