% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotluck}
\alias{plotluck}
\title{"I'm feeling lucky" for ggplot}
\usage{
plotluck(data, x, y = NULL, z = NULL, w = NULL,
  opts = plotluck.options(), ...)
}
\arguments{
\item{data}{a data frame}

\item{x,y,z}{column names}

\item{w}{weight column (optional)}

\item{opts}{a named list of options (optional)}

\item{...}{additional parameters will be passed to the respective geom_* objects}
}
\value{
a ggplot object
}
\description{
\code{plotluck} examines one, two, or three variables and creates, based on
their characteristics, a scatter, box, bar, density, hex or spine plot, or a
heat map. It also automates handling of observation weights, log-scaling of
axes, reordering of factor levels, and overlays of smoothing curves and
median lines.
}
\section{Determining the type of plot}{
 One-dimensional plots (i.e., when
  \code{y} and \code{z} are \code{NULL}) are either density plots (for
  numeric variables) or bar plots (for factors). In the former case, if the
  number of data points is not sufficient (as determined by the option
  \code{min.points.density}) a strip chart (1D scatter plot) is drawn.

  The following table summarizes the general heuristics for two dimensions
  (\code{x, y} are not \code{NULL}):

\tabular{lll}{
\strong{x}\tab \strong{y}\tab \strong{type}\cr
num\tab num\tab hexbin\cr
num\tab fact\tab density (histogram)\cr
fact\tab num\tab violin (box-and-whisker)\cr
fact\tab fact\tab spine\cr
}

In the first three cases, if the number of data points is not sufficient (as
determined by the options \code{min.points.hex}, \code{min.points.density},
or \code{min.points.box}), the plot switches to a scatter plot.

Density plots come with an overlaid vertical (weighted) median line; numeric scatter
plots with a smoothing line incuding standard deviations.
For three dimensions, the algorithm first tries sue coloring to fit the
\code{z} dimension into a single graph. If this is not possible (e.g., in the
case of the hexbin plot, which encodes the number of points in a bin by the
color; or as determined by the options \code{max.colors.scatter},
\code{max.colors.density}, \code{max.colors.box}, or \code{max.colors.bar}),
facetting is used, in a direction (i.e., horizontally or vertically subject
to \code{max.facets.row} and \code{max.facets.column})  that facilitates
comparing the particular type of plot. To this end, a numerical \code{z}
variable is discretized into \code{discretize.intervals.z} many intervals of
equal count or weight.

There are three notable exceptions from this rule:
\itemize{
\item If all three variables are categorical, a spine plot is drawn.
\item If the data is "grid-like", a heat map is drawn with the mean value of \code{z} providing the color.
The data is considered grid-like if
\itemize{
\item \code{x} and \code{y} are numeric or ordered factors, and \code{z} is numeric.
\code{x} and \code{y} have at least \code{min.size.grid.heat} distinct values each.
\item A sufficient fraction (as determined by option
\code{min.coverage.heat}) of the \code{(x,y)} grid points has at least one
data point.
}
\item If each factor combination occurs at most once in the data set, we resort to
bar plots.
}
}

\section{Reordering of factor levels}{
 To better illustrate the relation
 between an independent factor variable and a dependent numerical variable
 (or an ordered factor), factor levels are reordered according to the
 (weighted) mean of the dependent variable. Typically, the dependent variable
 is \code{y}, but it is \code{z} for spine plots and heat maps. If there is
 no dependent numeric variable, ordering proceeds by frequency.

 \code{plotluck} respects \emph{ordered} factors, so to prevent undesirable
 reordering, set the data type to \code{ordered}.
}

\section{Instance weights}{
 Argument \code{w} allows to specify weights or
 frequency counts for each row of data. All plots and summary lines take this
 weight into account (well, unfortunately all except hexbin, as the
 underlying implementation does not support it. We could approximate it by
 resampling with replacement, but this seems a little crude).

 In scatter plots, weights are indicated by a shaded disk whose area is
 proportional to the weight.

 There is some redundancy with regard to \emph{repeated points}, since they
 are conceptually identical to unique points with an associated frequency
 weight. By default, the option \code{convert.duplicates.to.weights} is
 switched on, which does exactly this. If it is \code{FALSE}, horizontal and
 vertical jittering will be applied if the number of duplicated points
 exceeds \code{min.points.jitter}. The amount of jittering can be controlled
 with \code{width.jitter}, \code{height.jitter}, and \code{factor.jitter}.
}

\section{Axis scaling}{
 \code{plotluck} supports logarithmic and log-modulus
 axis scaling. log-modulus is considered if values are both positive and
 negative; in this case, the transform function is \code{f(x) = sign(x) *
 log(1+abs(x))}.

 The decision whether to apply scaling is based on the proportion of total
 display range that is occupied by the 'core' region of the distribution
 (between the lower and upper quartiles); namely, if the transform could
 enlarge this region by a factor of at least \code{trans.log.thresh}.
}

\section{Missing values}{
 Typically, missing (\code{NA}) values in factors are
 ignored. If you want to make these explicit, make it a separate level using
 the \code{exclude} option in \code{\link{factor}}. In this case,
 \code{plotluck} will respect these levels. You can also explicitly set the
 exclude handling using the option \code{exclude.factor}, with the same
 semantics as in function \code{factor()}.
}

\section{Sampling}{
 For very large data sets, plots will take a very long time
 (or crash R). \code{plotluck} has a built-in stop-gap: If the data comprises
 more than \code{max.sample.rows}, it will be sampled down to that size
 (taking weights into account, if supplied).
}

\section{Factor preprocessing}{
 Frequently, when numeric variables have very
 few values despite sufficient data size, it helps to treat these values as
 the levels of a factor; this is governed by option
 \code{few.unique.as.factor}.

 If an unordered factor has too many levels, plots can get messy. In this
 case, only the \code{max.factor.levels} most frequent ones are retained,
 while the rest is binned together in a new default level.
}

\section{Column name matching}{
 Column names \code{x,y,z,w} are matched by
 unique prefix, and ignoring case.
}

\section{Remarks on the choice of plot types}{
 By default, \code{plotluck} uses
 violin and density plots in place of the more traditional box-and-whisker
 plots and histograms. The author feels that these convey the shape of a
 distribution better: in the former case, summary statistics like mean and
 quantiles are less useful if the distribution is not unimodal. A wrong
 choice of the number of bins of a histogram can create misleading artifacts.

 Nevertheless, the defaults can be overriden by changing options
 \code{use.geom.violin} and \code{use.geom.density}.

 Pie charts are not supported, as their flaws have been widely documented;
 see e.g., \code{\link{pie}}.

 Cleveland's Dot Plots (\code{\link{dotchart}}) can be produced as a special
 case of scatter plots.

 With real-world data (as opposed to smooth mathematical functions),
 three-dimensional scatter, surface, or contour plots can often be hard to
 read if the shape of the distribution is not suitable, data coverage is
 uneven, or if the perspective is not carefully chosen depending on the data.
 Therefore, we have refrained from incorporating them.
}

\section{Remarks on the use of options}{
 For completeness, we have included the
 description of option parameters in the current help page. However, the
 tenet of this function is to be usable "out-of-the-box", with no or very
 little manual tweaking required. If you find yourself needing to change
 option values repeatedly or find the presets to be suboptimal, please
 contact the author.
}

\section{What \code{plotluck} does not}{
 This function is designed for generic
 out-of-the-box plotting, and not suitable to produce more specialized types
 of plots that arise in specific application domains (e.g., association,
 stem-and-leaf, star plots, geo maps, etc). It is restricted to at most three
 variables. Parallel plots with variables on different scales (such as time
 series of multiple related signals) are not supported.
}
\examples{
#1D
#density plot
data(iris)
plotluck(iris, Petal.Length)

 # bar plot
data(movies, package='ggplot2')
plotluck(movies, mpaa)

#2D
# scatter plot
plotluck(iris, Petal.Length, Petal.Width)
plotluck(movies,votes,rating)

# density
plotluck(movies, rating, mpaa)         # using facets

data(diamonds, package='ggplot2')
plotluck(diamonds, price, cut)

plotluck(iris, Sepal.Length, Species)  # using colors

# box/violin plot
plotluck(movies, mpaa, rating)
plotluck(movies, Documentary, budget) # implicit conversion of binary variables

# spine plot
data(Titanic)
plotluck(as.data.frame(Titanic), Class, Survived, w=Freq)

data(occupationalStatus)
df <- as.data.frame(occupationalStatus)
df$origin <- ordered(df$origin)
df$destination <- ordered(df$destination)
plotluck(df, origin, destination, w=Freq)

# 3D
# heat map
plotluck(diamonds, cut, color, price)

# scatter plat
plotluck(movies, length, rating, mpaa)             # using facets
plotluck(iris, Petal.Length, Petal.Width, Species) # using colors

# density plot
plotluck(diamonds, price, cut, color)

# bar plot
plotluck(movies, mpaa, rating, Action) # grouped

# spine plot
plotluck(as.data.frame(Titanic), Class, Sex, Survived, w=Freq)
}
\seealso{
\code{\link{ggplot}}, \code{\link{plotluck.options}}, \code{\link{plotluck.multi}}
}
\concept{
automation

exploratory data analyis

ggplot

ggplot2

heat map

density plot

violin plot

hexbin

histogram

bar plot

box plot

spine plot

scatter plot

heat map
}
\keyword{aplot,}
\keyword{dplot}
\keyword{hplot,}

