% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotluck.R
\name{plotluck.options}
\alias{plotluck.options}
\title{Create options structure for \code{plotluck}}
\usage{
plotluck.options(...)
}
\arguments{
\item{...}{parameters to override default settings}
}
\value{
a named list of options, usable by function \code{plotluck}

 \code{plotluck} accepts a list of options to modify its behavior. Calling
 \code{plotluck.options} without arguments produces a list with the default
 values. Specifying any number of attribute/value pairs overrides these
 values selectively.

 Available options and their meaning is described inline in the documentation
 of \code{link{plotluck}}.
}
\description{
Create options structure for \code{plotluck}
}
\note{
\code{plotluck}'s aim is to provide a function that is usable
 "out-of-the-box", with no or very little manual tweaking. If you find
 yourself needing to change option values repeatedly or find the presets to
 be suboptimal, please contact the author.
}
\section{Shortcut options}{

\itemize{
\item By default, hex, density, box, and violin plots revert to
scatter plots for low number of points. \code{prefer.scatter=FALSE} (resp.
\code{TRUE}) sets \code{min.points.hex}, \code{min.points.density}, and
\code{min.points.box} to zero (resp. \code{Inf}).
\item If variable \code{z} is specified and the plot type, as well as its
number of levels, allows it, coloring is used to for representation in a single
plot; otherwise, \code{plotluck} reverts to faceting. Users can influence this
behavior by setting \code{prefer.color.for.z=TRUE} (resp. \code{FALSE}), which
causes \code{max.colors.scatter}, \code{max.colors.density},
\code{max.colors.box}, and \code{max.colors.bar} to be \code{Inf} (resp. 0).
}
}
\examples{
# list all default options
plotluck.options()

data(iris)
# default with violin plot
plotluck(iris, Species, Petal.Length)

# use box-and-whiskers plot instead
plotluck(iris, Species, Petal.Length, opts=plotluck.options(use.geom.violin=FALSE))

}
\seealso{
\code{\link{plotluck}}
}

